move manifest under pacman
